Avance

+++ src/main/java/com/bookstore/vehicleWeb/repository/VehicleRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface VehicleRepository extends JpaRepository<Vehicle,Double> {
+public interface VehicleRepository extends JpaRepository<Vehicle,Long> {
 
     List<Vehicle> findByBrand(String brand);
 }
